+++ telnetlib3/stream_writer.py
@@ -18,6 +18,7 @@ if TYPE_CHECKING:  # pragma: no cover
 
 # local
 from . import slc
+from .accessories import TRACE, hexdump
 from .mud import (
     zmp_decode,
     atcp_decode,
@@ -891,6 +892,8 @@ class TelnetWriter:
         assert isinstance(buf, (bytes, bytearray)), buf
         assert buf and buf.startswith(IAC), buf
         if not self.is_closing():
+            if self.log.isEnabledFor(TRACE):
+                self.log.log(TRACE, "send IAC %d bytes\n%s", len(buf), hexdump(buf, prefix=">>  "))
             self._transport.write(buf)
             if hasattr(self._protocol, "_tx_bytes"):
                 self._protocol._tx_bytes += len(buf)
@@ -2113,6 +2116,10 @@ class TelnetWriter:
                 # greater than 127, but it was removed for performance.
                 buf = self._escape_iac(buf)
 
+            if self.log.isEnabledFor(TRACE):
+                self.log.log(
+                    TRACE, "send %d bytes\n%s", len(buf), hexdump(buf, prefix=">>  ")
+                )
             self._transport.write(buf)
             if hasattr(self._protocol, "_tx_bytes"):
                 self._protocol._tx_bytes += len(buf)